Migrate the YAML serializer onto the shared serialization core (C2)
YamlSerializer now runs on the same architecture as its Bencode/TOML
siblings instead of its bespoke reflection walkers:

- Bodu.Text.Yaml compiles the shared Bodu.Text.Serialization/shared/**
  source under a new YAML symbol: the MetadataResolver/TypeMetadata/
  PropertyMetadata trio (compiled accessors, slot-indexed buffers, the
  full attribute family) and the structural converter factories
  (Nullable/Dictionary/Collection/Object). The token-strict shared
  scalar/enum/node converters and the engine seam are excluded — YAML's
  implicit typing coerces across scalar kinds, which those cannot
  express.
- MetadataResolver gains #if YAML branches preserving YAML's pinned
  contracts: duplicate wire names throw InvalidOperationException,
  extension-data misuse throws YamlSerializationException, and the
  extension-data member type is IDictionary<string, object?>.
- New format-local converters (Text.Yaml.Serialization.Converters)
  reproduce the retired Bind*/WriteValue semantics verbatim: cross-kind
  scalar coercion, NumberHandling float-to-integer policy, ignoreCase
  enum names, WriteEnumsAsStrings, decimal-as-exact-text, the dynamic
  object graph, runtime-type polymorphic writes, null members writing
  the null scalar, Hashtable-style untyped dictionaries with
  string-collision rejection, and the dotted failure paths. Null
  handling and the nesting-depth ceiling live at the single
  WriteAsObject dispatch seam, matching the walker's per-value order.
- YamlSerializerOptions adopts the sibling pipeline: a freeze-aware
  IList<YamlConverter> ConverterList, the non-null GetConverter
  resolution (converter attribute -> user converters -> defaults) with
  a concurrent cache, GetTypeMetadata, and InstantiateConverter. New
  public YamlConverterFactory completes the converter model.
- Utf8YamlWriter.WriteInt64 is renamed WriteInteger (sibling parity;
  the one deliberate public rename) and Utf8YamlReader gains Skip().
- Retired: YamlTypeBinding, YamlMemberInfo, YamlConverterCollection,
  and the YamlSerializer.Read.Collections walkers (net -1,000 lines of
  bespoke binding).

Adopted shared behaviors (previously absent, aligned with siblings):
constructor parameters without defaults are required members; concrete
non-IList collections (HashSet, Queue, Stack) and concrete custom
dictionaries now materialize correctly; NamingPolicyAttribute,
type-level UnmappedMemberHandling/ObjectCreationHandling, and
ConverterAttribute are honored.

Gates: YAML regression 1,429/1,429, Bencode 940/940, TOML 4,168/4,168.

+++ Bodu.Text.Yaml/src/Text.Yaml.Writer/Utf8YamlWriter.cs
@@ -152,7 +152,7 @@ public ref struct Utf8YamlWriter
     /// Writes an integer scalar value.
     /// </summary>
     /// <param name="value">The integer value.</param>
-    public void WriteInt64(long value) => EmitScalar(value.ToString(CultureInfo.InvariantCulture));
+    public void WriteInteger(long value) => EmitScalar(value.ToString(CultureInfo.InvariantCulture));
 
     /// <summary>
     /// Writes a floating-point scalar value.
